requirements.txt
./packages/azure_ai_projects-1.0.0b1-py3-none-any.whl # azure AI projects package
azure-ai-inference[prompts]
azure-identity
azure-monitor-opentelemetry
azure-search-documents
python-dotenv
openai
pandas
azure-ai-evaluation[remote]
